Reconfigure OpenSSL to allow engine and compression.
Some parts of Android such as the keystore engine (engine) and Chromium (compression) need these features.
Allow these configurations for now (for testing), but the usages need to be reviewed carefully
for FIPS compliance. We might end up changing the usages to not need engine support and
revert this change.

Change-Id: I321a17d9694a20c6628abaf7c956aaa296feb63c